Minor fixes
(cherry picked from commit 62021d238a0c38401b50211d59e72e169399f444)

+++ src/main/java/com/github/technus/tectech/mechanics/elementalMatter/definitions/complex/EMHadronDefinition.java
@@ -101,7 +101,7 @@ public final class EMHadronDefinition extends EMComplexTemplate {//TODO Optimize
         this.amount = amount;
         this.charge = charge;
         this.type = containsAnti ? (byte) -type : (byte) type;
-        long mult = this.getAmount() * this.getAmount() * (this.getAmount() - 1);
+        long mult = (long) this.getAmount() * this.getAmount() * (this.getAmount() - 1);
         mass = mass * 5.543D * mult;//yes it becomes heavier
 
         if (mass == protonMass && this.getAmount() == 3) {
